debug-assignment/blood-test-analyser-debug/requirements.txt
# Core FastAPI and web framework
fastapi==0.110.3
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.4

# AI and ML
crewai==0.130.0
langchain==0.3.26
langchain-openai==0.3.27
langchain-community==0.3.26
langchain-google-genai==2.1.5
openai==1.93.0
google-generativeai==0.5.4

# PDF processing - Optimized versions
pypdf==5.6.1
PyPDF2==3.0.1
pdfplumber==0.11.7

# Database - Using PyMySQL for Windows compatibility
sqlalchemy==2.0.38
pymysql==1.1.0
alembic==1.14.1

# Task queue and caching
celery==5.3.6
redis==5.2.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Utilities
pydantic==2.10.6
pydantic-settings==2.8.0
python-dateutil==2.9.0

# Development and testing
pytest==8.2.1
pytest-asyncio==0.24.0

# Performance and monitoring
psutil==7.0.0
memory-profiler==0.61.0

# Additional dependencies for better performance
aiofiles==24.1.0
httpx==0.28.0
tenacity==8.2.3

# Keep crewai at version 0.130.0
# Other package versions are flexible and can be changed
# Only change crewai version if there are critical dependency conflicts that cannot be resolved by other means
crewai-tools==0.47.1
google-ai-generativelanguage==0.6.4
google-api-core==2.10.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.53.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
jsonschema==4.22.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
pandas==2.2.2
pillow==10.3.0
pip==24.0
protobuf==4.25.3
pydantic_core==2.8.0